feat(c): change sum variable type to float for precision
- Changed the type of the variable 'sum' from int to float in ten-average.c
- This change allows for more precise calculations of the average when dealing with decimal values.

This update enhances the accuracy of the average calculation functionality, ensuring that the program can handle a wider range of input values effectively.

+++ c/ten-average.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int i;
-int sum;
+float sum;
 float average;
 
 void main() {
